project/requirements.txt
# --- Core framework ---
Django==5.2.10

# --- REST API ---
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# --- Production basics ---
gunicorn==23.0.0
whitenoise==6.10.0
django-environ==0.12.0
django-cors-headers==4.7.0

# --- Database driver (Postgres) ---
psycopg[binary]==3.2.9

# --- External API calls (RapidAPI) ---
requests==2.32.3
